chore(benchmarks): add build tags to make corpus benchmarks opt-in
Corpus benchmarks (BenchmarkCorpus_*) now require -tags=corpus to run.
This prevents accidental memory exhaustion when running `go test -bench=.`

Changes:
- Create corpus_bench_test.go files with //go:build corpus tag
- Move corpus benchmarks from original test files to tagged files
- Update Makefile bench-corpus target to use -tags=corpus
- Simplify bench-release target (no longer needs -skip flag)
- Update backfill-benchmarks.sh with documentation comment

Run corpus benchmarks with: make bench-corpus
Or directly: go test -tags=corpus -bench=BenchmarkCorpus ./...
