+++ fixer/fixer_test.go
@@ -647,27 +647,5 @@ paths:
 	}
 }
 
-// BenchmarkCorpus_Fix benchmarks fixing a real corpus spec
-func BenchmarkCorpus_Fix(b *testing.B) {
-	spec := corpusutil.GetByName("Asana")
-	if spec == nil {
-		b.Skip("Asana spec not found")
-	}
-	if !spec.IsAvailable() {
-		b.Skipf("Corpus file %s not cached", spec.Filename)
-	}
-
-	p := parser.New()
-	parseResult, err := p.Parse(spec.GetLocalPath())
-	if err != nil {
-		b.Fatal(err)
-	}
-
-	for b.Loop() {
-		f := New()
-		_, err := f.FixParsed(*parseResult)
-		if err != nil {
-			b.Fatal(err)
-		}
-	}
-}
+// Note: BenchmarkCorpus_Fix has been moved to corpus_bench_test.go
+// Run with: go test -tags=corpus -bench=BenchmarkCorpus ./fixer/...
